updated the utils section and modularised the codebase
